@@ -16,15 +16,17 @@
     "src/index.ts"
   ],
   "devDependencies": {
-    "@types/mocha": "^5.2.5",
-    "@types/node": "^10.12.18",
-    "@types/webpack": "^4.4.24",
-    "foy": "^0.1.12",
+    "@types/mocha": "^8.2.2",
+    "@types/node": "^14.14.35",
+    "foy": "^0.2.3",
     "mocha": "^5.2.0",
     "ts-node": "^8.0.2",
     "tsconfig-paths": "^3.7.0",
-    "typescript": "^3.2.4",
-    "webpack": "^4.29.0"
+    "typescript": "^4.2.3",
+    "webpack": "^5.27.2"
+  },
+  "peerDependencies": {
+    "webpack": "^5"
   },
   "dependencies": {
     "ejs": "^2.6.1"
